LPARDesign
The LPAR Design tool assists you in planning the LPAR layout of your Central Processor Complexes. The tool allows you
to specify all partitions, the number of logical processors and their weights. If you run your system in Hiperdispatch mode it
also assist you in displaying the number of high, medium and low processors as a result of your definition. This will help
you to easily identify definition errors. In addition offload processors like zIIPs and zAAPs are also supported.
You can upload or download the results from / to a zPCR study. To install the tool, download the LPAR Design Tool and
unzip it to your workstation. The package consists of the tool (a Microsoft Excel spreadsheet) and the associated user's
guide in the PDF format.